move test util files to test folder instead of util folder
