+++ exports/src/utils.ts
@@ -1,4 +1,4 @@
 export * from '../../src/utils/inputSanitizer';
 export * from '../../src/utils/misc';
-export * from '../../src/utils/testUtils';
-export * from '../../src/utils/setupTestDB';
+export * from '../../src/test/utils/testUtils';
+export * from '../../src/test/utils/setupTestDB';
