+++ src/controller/authController.test.ts
@@ -1,5 +1,5 @@
-import { setUpInMemDB } from '../utils/setupTestDB';
-import { wait, expectString, expectDate } from '../utils/testUtils';
+import { setUpInMemDB } from '../test/utils/setupTestDB';
+import { wait, expectString, expectDate } from '../test/utils/testUtils';
 import { con_auth_register, con_auth_login, con_auth_refresh, con_auth_logout } from './authController';
 import { InputError, AuthError, ConflictError } from '../error/AppError';
