+++ src/services/authServices.test.ts
@@ -1,5 +1,6 @@
 import mongoose, { Types } from 'mongoose';
-import { setUpInMemDB } from '../utils/setupTestDB';
+import { setUpInMemDB } from '../test/utils/setupTestDB';
+import { expectMongooseDoc } from '../test/utils/testUtils';
 import {
   ser_createUser,
   ser_findUserViaEmail, ser_findUserViaRT, ser_findUserViaId,
@@ -7,7 +8,6 @@ import {
 } from './authServices';
 import { ConflictError } from '../error/AppError';
 import { UserModel } from '../models/userModel';
-import { expectMongooseDoc } from '../utils/testUtils';
 import { hashValue } from '../utils/hash';
 
 setUpInMemDB();
