+++ src/test/integration/tokenInvalidation.test.ts
@@ -1,7 +1,7 @@
-import { setUpInMemDB } from '../../utils/setupTestDB';
+import { setUpInMemDB } from '../../test/utils/setupTestDB';
+import { wait } from '../../test/utils/testUtils';
 import { con_auth_register, con_auth_login, con_auth_refresh, con_auth_logout } from '../../controller/authController';
 import { AuthError } from '../../error/AppError';
-import { wait } from '../../utils/testUtils';
 
 setUpInMemDB();
