+++ src/utils/inputSanitizer.test.ts
@@ -8,7 +8,7 @@ import {
   sanitizeDefaultSprintColumns,
   sanitizeObjectId
 } from './inputSanitizer';
-import { testInvalidStringInputs } from './testUtils';
+import { testInvalidStringInputs } from '../test/utils/testUtils';
 import { EMAIL_MAX_LEN, PW_MAX_LEN, TITLE_MIN_LEN, TITLE_MAX_LEN, DESC_MAX_LEN, SPRINT_COLS_MAX } from '../consts';
 import { InputError } from '../error/AppError';
